requirements.txt
anyio==4.4.0
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
packaging==24.0
pyTelegramBotAPI==4.18.1
python-dotenv==1.0.1
python-telegram-bot==21.2
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.0
urllib3==2.2.1
